crash (4.0-8.11-1) unstable; urgency=low

  * New upstream version 4.0-8.11 (see
    http://people.redhat.com/anderson/crash.changelog)
  - Fix backtrace when offline CPUs doe not have NT_PRSTATUS notes
  - Fix "dev" command for 2.6.26 later kernels.
  - Fix segfaul when running "foreach bt" on active live system.
  - Fix to read LKCD V8 x86 dumpfile headers
  - Fix bt when NMI was issued ot non-crashing amd64 CPU.
  - Fix bt on amd64 when transitioning from the IRQ stack back to the process
  stack on >= 2.6.29 kernels.
  - Enhance "runq" to show current task in each cpu's runqueue. 
  - Fix mem leak when running on live systems.
  - Support dwarf 3 vmlinux debuginfo format
  - Allow any order for system map and compressed vmlinux file on the cmd
  line.
  - Fix seg fault when vmcore and system map do not match
  - Adjust virt and phys memory definitions for >= 2.6.31
  - Fix for "kmem -[sS]" when using CONFIG_SLUB and there is an invalid link.

+++ lkcd_dump_v8.h
@@ -445,9 +445,9 @@ typedef struct _dump_header_asm_s {
  */
 typedef struct dump_CPU_info_s {
 	struct pt_regs	dha_smp_regs;
-	uint64_t	dha_smp_current_task;
-	uint64_t	dha_stack;
-	uint64_t	dha_stack_ptr;
+	uint32_t	dha_smp_current_task;
+	uint32_t	dha_stack;
+	uint32_t	dha_stack_ptr;
 } __attribute__ ((packed)) dump_CPU_info_t;
